aiagent/requirements.txt
# Core AI and ML dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web framework and API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# HTTP client and utilities
httpx>=0.24.0
aiofiles>=23.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3
